use parent block header for building the block (#117)
* use block header and add tests

* try somthing

* final changes

* remove reorg test

* remove unused stuuf

* update name

* revert change

* address comments and add mutex

* whitespace

* Apply suggestion from @Copilot


* fix logs

* fix format

* Refac

* fix

---------

+++ src/Nethermind.Arbitrum/Execution/ArbitrumBlockProducer.cs
@@ -54,7 +54,7 @@ namespace Nethermind.Arbitrum.Execution
         protected BlockHeader PrepareBlockHeader(BlockHeader parent, ArbitrumPayloadAttributes payloadAttributes, ArbosState arbosState)
         {
             long newBlockNumber = parent.Number + 1;
-            if (payloadAttributes.Number != (ulong)newBlockNumber)
+            if (payloadAttributes.Number != newBlockNumber)
                 throw new ArgumentException($"Wrong message number in digest, got {payloadAttributes.Number}, expected {newBlockNumber}");
 
             ulong timestamp = payloadAttributes?.MessageWithMetadata.Message.Header.Timestamp ?? UInt64.MinValue;
